+++ src/app/app.routing.ts
@@ -13,51 +13,100 @@ import { BravoMonacoEditorBaseDemo } from './demo/bravo.monaco.editor.base.demo/
 import { BravoDocxtemplater } from './components/bravo.docxtemplater/bravo.docxtemplater';
 import { BravoWebviewer } from './components/bravo.webviewer/bravo.webviewer';
 import { BravoTabGridLayoutDemo } from './demo/bravo.tab.grid.layout.demo/bravo.tab.grid.layout.demo';
+import { BravoFullScreen } from './layouts/bravo.full.screen/bravo.full.screen';
+import { BravoTools } from './layouts/bravo.tools/bravo.tools';
+import { BravoDemo } from './layouts/bravo.demo/bravo.demo';
 
 const routes: Routes = [
     {
-        path: 'bravo-tab-grid-layout-demo',
-        component: BravoTabGridLayoutDemo
+        path: 'full',
+        component: BravoFullScreen,
+        children: [
+            {
+                path: 'bravo-docxtemplater',
+                component: BravoDocxtemplater
+            },
+            {
+                path: 'bravo-toolbar',
+                component: BravoToolbar
+            },
+            {
+                path: 'bravo-picture-editor',
+                component: BravoPictureEditor
+            },
+            {
+                path: 'bravo-dialog-data-button',
+                component: BravoDialogDataButton
+            },
+            {
+                path: 'bravo-tab-grid-layout',
+                component: BravoTabGridLayout
+            },
+            {
+                path: 'bravo-picture-input-box',
+                component: BravoPictureInputBox
+            },
+            {
+                path: 'bravo-webviewer',
+                component: BravoWebviewer
+            }
+        ]
     },
     {
-        path: 'bravo-monaco-editor-base-demo',
-        component: BravoMonacoEditorBaseDemo
+        path: 'tools',
+        component: BravoTools,
+        children: [
+            {
+                path: 'bravo-docxtemplater',
+                component: BravoDocxtemplater
+            },
+            {
+                path: 'bravo-toolbar',
+                component: BravoToolbar
+            },
+            {
+                path: 'bravo-picture-editor',
+                component: BravoPictureEditor
+            },
+            {
+                path: 'bravo-dialog-data-button',
+                component: BravoDialogDataButton
+            },
+            {
+                path: 'bravo-tab-grid-layout',
+                component: BravoTabGridLayout
+            },
+            {
+                path: 'bravo-picture-input-box',
+                component: BravoPictureInputBox
+            },
+            {
+                path: 'bravo-webviewer',
+                component: BravoWebviewer
+            }
+        ]
     },
     {
-        path: 'bravo-monaco-editor-demo',
-        component: BravoMonacoEditorDemo
-    },
-    {
-        path: 'bravo-picture-editor-demo',
-        component: BravoPictureEditorDemo
-    },
-    {
-        path: 'bravo-docxtemplater',
-        component: BravoDocxtemplater
-    },
-    {
-        path: 'bravo-toolbar',
-        component: BravoToolbar
-    },
-    {
-        path: 'bravo-picture-editor',
-        component: BravoPictureEditor
-    },
-    {
-        path: 'bravo-dialog-data-button',
-        component: BravoDialogDataButton
-    },
-    {
-        path: 'bravo-tab-grid-layout',
-        component: BravoTabGridLayout
-    },
-    {
-        path: 'bravo-picture-input-box',
-        component: BravoPictureInputBox
-    },
-    {
-        path: 'bravo-webviewer',
-        component: BravoWebviewer
+        path: 'demo',
+        component: BravoDemo,
+        children: [
+            {
+                path: 'bravo-tab-grid-layout-demo',
+                component: BravoTabGridLayoutDemo
+            },
+            {
+                path: 'bravo-monaco-editor-base-demo',
+                component: BravoMonacoEditorBaseDemo
+            },
+            {
+                path: 'bravo-monaco-editor-demo',
+                component: BravoMonacoEditorDemo
+            },
+            {
+                path: 'bravo-picture-editor-demo',
+                component: BravoPictureEditorDemo
+            }
+        ]
     }
 ];
