+++ src/app/layouts/bravo.demo/bravo.demo.html
@@ -1,0 +1,3 @@
+<div class="demo">
+    <router-outlet></router-outlet>
+</div>
