+++ src/app/layouts/bravo.demo/bravo.demo.scss
@@ -1,0 +1,4 @@
+.demo {
+    width: 100vw;
+    height: 100vh;
+}
