+++ src/app/layouts/bravo.demo/bravo.demo.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'bravo-demo',
+    templateUrl: './bravo.demo.html',
+    styleUrls: ['./bravo.demo.scss']
+})
+export class BravoDemo implements OnInit {
+    constructor() {}
+
+    ngOnInit(): void {}
+}
