+++ src/app/layouts/bravo.full.screen/bravo.full.screen.html
@@ -1,0 +1,3 @@
+<div class="full">
+    <router-outlet></router-outlet>
+</div>
