+++ src/app/layouts/bravo.full.screen/bravo.full.screen.scss
@@ -1,0 +1,4 @@
+.full {
+    width: 100vw;
+    height: 100vh;
+}
