+++ src/app/layouts/bravo.full.screen/bravo.full.screen.ts
@@ -1,0 +1,12 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'bravo-full-screen',
+    templateUrl: './bravo.full.screen.html',
+    styleUrls: ['./bravo.full.screen.scss']
+})
+export class BravoFullScreen implements OnInit {
+    constructor() {}
+
+    ngOnInit(): void {}
+}
