+++ src/app/layouts/bravo.tools/bravo.tools.scss
@@ -1,0 +1,118 @@
+.bravo-tools {
+    display: flex;
+    flex-direction: column;
+    width: 100vw;
+    height: 100vh;
+
+    .header {
+        height: 47px;
+        width: 100%;
+        border-top: 3px solid #119489;
+        box-shadow: 0 1px 2px hsla(0, 0%, 0%, 0.05), 0 1px 4px hsla(0, 0%, 0%, 0.05), 0 2px 8px hsla(0, 0%, 0%, 0.05);
+
+        .tools {
+            display: flex;
+            flex-direction: row;
+            justify-content: space-between;
+            margin: 0 10px 0 10px;
+            height: 100%;
+            width: calc(100% - 20px);
+            overflow: auto;
+
+            .viewport {
+                display: flex;
+                flex-direction: row;
+                height: 100%;
+
+                .input {
+                    display: flex;
+                    flex-direction: row;
+                    align-items: center;
+                    width: 200px;
+                    height: 100%;
+                    margin-right: 20px;
+
+                    input {
+                        width: 100%;
+                        min-width: 60px;
+                        height: 30px;
+                        margin: 3px;
+                    }
+                }
+
+                .event {
+                    display: flex;
+                    flex-direction: row;
+                    align-items: center;
+                    height: 100%;
+
+                    button {
+                        width: 60px;
+                        height: 34px;
+                        margin: 3px;
+
+                        &:hover {
+                            cursor: pointer;
+                            background-color: #119489;
+                            border: none;
+                            border-radius: 2px;
+                            color: white;
+                        }
+                    }
+                }
+            }
+
+            .theme {
+                display: flex;
+                flex-direction: row;
+                align-items: center;
+
+                .background {
+                    min-width: 215px;
+                    height: 34px;
+                    margin-left: 20px;
+                }
+            }
+        }
+    }
+
+    .body {
+        margin: 10px;
+        width: calc(100% - 20px);
+        height: calc(100% - 70px);
+        overflow: auto;
+
+        .content {
+            width: 100%;
+            height: 100%;
+
+            .component {
+                width: 100%;
+                height: 100%;
+                overflow: hidden;
+                box-shadow: rgba(0, 0, 0, 0.1) 0px 4px 12px;
+            }
+        }
+    }
+
+
+}
+
+@media screen and (max-width: 958px) {
+    .bravo-tools {
+        .header {
+            height: 64px;
+
+            .tools {
+                margin: 0 5px 0 5px;
+                width: calc(100% - 10px);
+            }
+        }
+
+        .body {
+            margin: 5px;
+            width: calc(100% - 10px);
+            height: calc(100% - 60px);
+        }
+    }
+}
